requirements.txt
aiohttp
async-timeout
beautifulsoup4
pyyaml
pandas
numpy
tqdm
playwright
